new commit

+++ controllers/cookiecontroller.js
@@ -27,23 +27,23 @@ router.get ("/", (req, res) => {
         .catch(err => res.status(500).json({err: err}))
 })
 
-router.get('/:id', (req, res) => {
+router.get('/mycookies', (req, res) => {
     Cookie.findAll({ where:
-        { id: req.params.id }
+        { owner_id: req.user.id }
     })
         .then(cookies => res.status(200).json(cookies))
         .catch(err => res.status(500).json({ error: err}))
     })
 
 router.put('/:id', validateSession, (req,res) => {
-    Cookie.update(req.body.cookie, { where: { id: req.params.id }, returning: true
+    Cookie.update(req.body.cookie, { where: { id: req.params.id, owner_id: req.user.id }, returning: true
 })
     .then(cookie => res.status(200).json(cookie))
     .catch(err => res.status(500).json({err: err}))
 })
 
 router.delete('/:id', (req,res) => {
-    Cookie.destroy({ where: { id: req.params.id}
+    Cookie.destroy({ where: { id: req.params.id, owner_id: req.user.id}
 })
     .then(cookie => res.status(200).json(cookie))
     .catch(err => res.status(500).json({err: err}))
